Rewriting network requests to use urllib, because of inexplicable bugs
The main JewelData request was failing completely with a mysterious
parse error. Copying the same url and loading it in the browser worked
just fine, so I'm left to suspect that QNetworkRequest is truncating the
URL or something stupid. The stock urllib request works fine, so we're
switching to wrapping one of those in a simple QThread. Frankly, I like
it better this way anyway.