Workaround for PyQt terminating null bug
PyQt4 seems to have a weird bug where, when the JS string passed in
contains surrogate pairs (unicode chars that don't fit in a wchar, like
these: "𝟘𝟙𝟚𝟛𝟜𝟝𝟞𝟟𝟠𝟡"), those pairs are parsed correctly into single
Python characters, but an extra '\x00' character is appended to the end
of the string for each pair. JSON decoding chokes on those, so now we
strip them from the payload passed to arbiterInformSerialized.

TODO: Will we need a more general workaround for this in the future?
Could any other interfaces be affected?