Window positioning
The main window's rectangle is now recorded as a setting. The chat
window is positioned next to the main window, on the left if there's
enough room, but if the user moves it the application will save the new
position until it restarts. (This imitates the behavior of the Windows
version.)