Exit with sys.exit instead of application.quit
Sometimes strange things happen when the application exits by closing
the main window. About half the time, I see a segfault in the terminal
output. I've also seen the app live forever after the window closes,
burning 100% cpu on who knows what. Presumably this is because other
parts of the program are still making calls on some Qt objects that are
no longer valid. Since this is only coming up on application exit, the
simple thing to do is to just sys.exit() directly, instead of trying to
clean up Qt.