chat window alerts
The chat window pops up when a message is received, but it wasn't doing
anything if another message came in while the window was open in the
background. Here we add a simple call to Qt's alert() method. I tried
this in the past and found that it was buggy on Ubuntu, but it seems to
work well enough on Gnome, so I'm adding it back. Might need to refine
this later with platform-specific hacks.